requirements.txt
uvicorn
fastapi
streamlit
pandas
openai
pyyaml
-e .